- Fixed deletion path on Linux/UNIX systems - Removed comment and  minor improvements

+++ ChatSQL/model.py
@@ -116,14 +116,17 @@ class ModelDelete:
                     os.remove(file_path)
                     json_deleted = True 
 
-            for index_file in os.listdir(self.indexes_path + "\\" + os.path.splitext(file)[0]):
-                file_path = os.path.join(self.indexes_path, os.path.splitext(file)[0], index_file)
-                if os.path.isfile(file_path):
-                    os.remove(file_path)
-                    shutil.rmtree(os.path.join(self.indexes_path, os.path.splitext(file)[0]))
-                    index_deleted = True
+            filename_without_extension = os.path.splitext(file)[0]
+            indexes_directory = os.path.join(self.indexes_path, filename_without_extension)
+            if os.path.isdir(indexes_directory):
+                for index_file in os.listdir(indexes_directory):
+                    file_path = os.path.join(indexes_directory, index_file)
+                    if os.path.isfile(file_path):
+                        os.remove(file_path)
+                shutil.rmtree(indexes_directory)
+                index_deleted = True
 
-        if json_deleted == True and index_deleted == True:
+        if json_deleted and index_deleted:
             self.file_deleted = True
     
     def getEsitoFileEliminato(self):
